+++ src/test/java/net/snowflake/client/jdbc/ResultSetLatestIT.java
@@ -13,6 +13,7 @@ import java.sql.*;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
+import java.util.TimeZone;
 import java.util.regex.Pattern;
 import net.snowflake.client.category.TestCategoryResultSet;
 import net.snowflake.client.jdbc.telemetry.*;
@@ -299,5 +300,52 @@ public class ResultSetLatestIT extends ResultSet0IT {
 
   private byte[] floatToByteArray(float i) {
     return ByteBuffer.allocate(Float8Vector.TYPE_WIDTH).putDouble(0, i).array();
+  }
+
+  @Test
+  public void testGetWallclockTimeToString() throws SQLException {
+    // set up
+    TimeZone tzOriginal = TimeZone.getDefault();
+    TimeZone.setDefault(TimeZone.getTimeZone("PST"));
+    Connection con = init();
+    Statement statement = con.createStatement();
+    statement.execute("alter session set timezone='UTC'");
+    statement.execute("create or replace table timetable (coltime time, coltimestamp timestamp)");
+    String timeVal = "17:17:17";
+    String offsetTimeVal = "09:17:17";
+    String timeNanos = timeVal + ".54321";
+    statement.execute(
+        "insert into timetable values ('" + timeVal + "', '1970-01-01 " + timeVal + "')");
+    statement.execute(
+        "insert into timetable values ('" + timeNanos + "', '1970-01-01 " + timeNanos + "')");
+    // Test getTime with default behavior, JVM PST timezone (offset of 8 hours from UTC)
+    ResultSet rs = statement.executeQuery("select * from timetable");
+    rs.next();
+    assertEquals(offsetTimeVal, rs.getTime(1).toString()); // time is offset by 8 hours
+    assertEquals(offsetTimeVal, rs.getTime(2).toString()); // time is offset by 8 hours
+    rs.next();
+    assertEquals(
+        offsetTimeVal, rs.getTime(1).toString()); // time is offset by 8 hours, no nanos by default
+    assertEquals(
+        offsetTimeVal, rs.getTime(2).toString()); // time is offset by 8 hours, no nanos by default
+    // Test with new parameter, JVM PST timezone (offset of 8 hours from UTC)
+    statement.execute("alter session set JDBC_USE_WALLCLOCK_TIME=true");
+    rs = statement.executeQuery("select * from timetable");
+    rs.next();
+    assertEquals(
+        timeVal, rs.getTime(1).toString()); // time value matches inserted value; no offset!
+    assertEquals(
+        timeVal, rs.getTime(2).toString()); // timestamp value matches inserted value; no offset!
+    rs.next();
+    assertEquals(timeNanos, rs.getTime(1).toString()); // No offset! Nanos do appear by default
+    assertEquals(timeNanos, rs.getTime(2).toString()); // No offset! Nanos do appear by default
+
+    // tear down
+    TimeZone.setDefault(tzOriginal);
+    rs.close();
+    statement.execute("alter session unset timezone");
+    statement.execute("drop table if exists timetable");
+    statement.close();
+    con.close();
   }
 }
